# services/billing/openapi.yaml
openapi: 3.0.3
info:
  title: Billing Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /api/v1/billing
paths:
...
  /tokens/reserve:
    post:
      operationId: reserveTokens
      summary: Reserve tokens for a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: integer }
                taskId: { type: string }
              required: [amount]
      responses:
        '202': { description: Reserved }
        '400': { description: Invalid input }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...